requirements.txt
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.0.5
django-adminlte-3==0.1.6
django-appconf==1.0.5
django-cors-headers==3.12.0
django-dotenv==1.4.2
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.4
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.9.1
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.13
